requirements.txt
fastapi==0.104.1
starlette==0.27.0
uvicorn==0.24.0
faster-whisper==0.10.0
deep-translator==1.11.4
yt-dlp==2025.5.22
ytmusicapi==1.3.2
ffmpeg-python==0.2.0
colorthief==0.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-multipart==0.0.6
requests==2.31.0
Pillow==10.1.0
aiofiles==23.2.1
httpx>=0.23.0
